Remove guard check from gem-build-dev-lxc resource
The git resource is always ensuring the dev-lxc source
is up to date. With the guard check removed the
gem-build-dev-lxc resource ensures the gem is rebuilt
on each run.  Now when we want to actually install the
latest we just need to run `gem uninstall -x dev-lxc`
and then run `vagrant provision`.