Configure dnsmasq to resolve the .lxc domain
Without this configured you can end up disrupting the VM's networking.
If you try to have dnsmasq resolve a .lxc FQDN's that isn't registered in
DHCP it hands off the resolution to the Ubuntu 13.10 host resolver.
The host VM is configured with --natdnsresolver1 = "on" so it hands off
the resolution back to the OS X host which should have /etc/resolver/lxc
which points right back to the dnsmasq DNS server inside the VM.